feat: useDropModal 커스텀 훅으로 분리
- useDropModal로 모달 기능 분리
- Noti 모달과 구분하기 위해 Drop 모달로 일부 리네이밍

#66
