+++ src/components/rightSide/user/UserInfo.tsx
@@ -1,8 +1,8 @@
-import { useCallback, useContext } from "react";
-import { UserDropContext } from "hooks/UserDropContext";
+import { useContext } from "react";
 import { ProfileContext } from "hooks/ProfileContext";
 import UserDropMenu from "./UserDropMenu";
 import * as S from "./style";
+import useDropModal from "hooks/useDropModal";
 
 export default function UserInfo(props: {
   listOf?: string;
@@ -11,17 +11,10 @@ export default function UserInfo(props: {
   subLine: string;
 }) {
   const setProfileUser = useContext(ProfileContext);
-  const dropped = useContext(UserDropContext);
-  let droppedUser = "";
-  if (props.listOf) droppedUser = props.listOf + props.username;
-
-  const onClose = useCallback(() => {
-    dropped?.userSet("");
-  }, []);
-
-  function onOpen() {
-    dropped?.userSet(droppedUser);
-  }
+  const { onDropOpen, onDropClose, dropIsOpen } = useDropModal({
+    listOf: props.listOf,
+    username: props.username,
+  });
 
   return (
     <>
@@ -39,15 +32,16 @@ export default function UserInfo(props: {
         <br />
         {props.subLine}
       </S.UserInfoText>
-      {droppedUser && <S.KebabIcon onClick={onOpen} />}
-      {droppedUser && dropped?.user === droppedUser && <UserDropMenu onClose={onClose} />}
-      {!props.listOf && (
+      {props.listOf ? (
+        <S.KebabIcon onClick={onDropOpen} />
+      ) : (
         <>
           {/* TODO: 새 초대가 있는 경우/없는 경우 조건 추가 */}
           <S.EmptyInviteIcon />
           <S.NewInviteIcon />
         </>
       )}
+      {dropIsOpen && <UserDropMenu onClose={onDropClose} />}
     </>
   );
 }
